# src/main/resources/swagger/api.yml
openapi: 3.0.3
info:
  title: Baseline Seed Admin API
  version: 1.0.0
  description: >
    Administrative endpoints for resetting legacy demo data and applying the curated
    trading baseline required for the RNExchange MVP (exchanges, broker, instruments,
    contracts, holidays, margin rules, and role mappings).
servers:
  - url: https://api.rnexchange.local
    description: Local development/staging
  - url: https://api.rnexchange.prod
    description: Production
paths:
  /api/admin/baseline-seed/run:
    post:
      summary: Trigger baseline cleanup and reseed job
      description: >
        Truncates non-system tables, replays Liquibase migrations, and applies the
        `baseline` context changelog with deterministic seed data. Responds asynchronously
        with a job descriptor. Requires `EXCHANGE_OPERATOR` authority.
      operationId: runBaselineSeed
      tags:
        - baseline-seed
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaselineSeedRequest'
      responses:
        '202':
          description: Seed job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaselineSeedJob'
        '409':
          description: A seed job is already running and cannot be restarted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Caller lacks `EXCHANGE_OPERATOR` authority
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/admin/baseline-seed/status/{jobId}:
    get:
      summary: Fetch status of a baseline seed job
      description: Polls the lifecycle of a submitted seed job, including metrics and error messages.
      operationId: getBaselineSeedStatus
      tags:
        - baseline-seed
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Current job status payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaselineSeedJob'
        '404':
          description: Job ID not found or expired from registry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Caller lacks `EXCHANGE_OPERATOR` authority
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT issued to an `EXCHANGE_OPERATOR` user'
    basic:
      type: http
      description: Basic Authentication
      scheme: basic
  schemas:
    BaselineSeedRequest:
      type: object
      properties:
        force:
          type: boolean
          default: false
          description: Force cleanup even if tables appear empty.
        contexts:
          type: array
          items:
            type: string
          description: Optional Liquibase contexts to include (defaults to `baseline`).
    BaselineSeedJob:
      type: object
      required:
        - jobId
        - status
        - submittedAt
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - PENDING
            - RUNNING
            - COMPLETED
            - FAILED
        submittedAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        durationMs:
          type: integer
          format: int64
          nullable: true
        metrics:
          type: object
          additionalProperties:
            type: string
          description: Key metrics (e.g., seeded counts per entity).
        errors:
          type: array
          items:
            type: string
          description: Validation or processing errors when status is `FAILED`.
    ProblemDetail:
      type: object
      required:
        - status
        - title
      properties:
        status:
          type: integer
          example: 403
        title:
          type: string
          example: Access denied
        detail:
          type: string
          description: Human-readable explanation, suitable for logs/UI.
security:
  - bearerAuth: []
  - basic: []